requirements.txt
gradio>=5.0.0
pandas>=2.0.0
httpx>=0.27.0
python-dotenv>=1.0.0
mcp>=1.0.0
requests>=2.31.0
